+++ src/web/assets/cp/src/js/CategoryIndex.js
@@ -132,15 +132,14 @@ Craft.CategoryIndex = Craft.BaseElementIndex.extend({
         // Update the URL if we're on the Categories index
         // ---------------------------------------------------------------------
 
-        if (this.settings.context === 'index' && typeof history !== 'undefined') {
+        if (this.settings.context === 'index') {
             let uri = 'categories';
 
             if (selectedSourceHandle) {
                 uri += '/' + selectedSourceHandle;
             }
 
-            const url = Craft.getUrl(uri, document.location.search + document.location.hash);
-            history.replaceState({}, '', url);
+            Craft.setPath(uri);
         }
     },